refactors PokemonList component
