+++ src/components/Homepage/PokemonList/StyledPokemonList.ts
@@ -1,12 +1,12 @@
-import styled, { css } from 'styled-components'
-import Box from '../../Box'
+import styled, { css } from 'styled-components';
+import Box from '@/components/Box';
 
 const Container = styled(Box)`
   ${({ theme }) => css`
     background-color: ${theme.homepage.pokemonList.backgroundColor};
     color: ${theme.homepage.pokemonList.color};
   `}
-`
+`;
 
 const SelectContainer = styled(Box)`
   div {
@@ -20,6 +20,6 @@ const SelectContainer = styled(Box)`
   div:nth-of-type(1) {
     margin-right: 1.5rem;
   }
-`
+`;
 
-export { Container, SelectContainer }
+export { Container, SelectContainer };
